pyproject.toml
[project]
name = "stageflow-production-testers"
version = "0.1.0"
description = "Stageflow stress-testing harness with reusable LLM/STT/TTS components"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "stageflow-core>=0.5.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools]
packages = ["components", "scripts"]
